FUN-27 Split checkout page into components
