+++ client/components/Atoms/form/TheSelect.vue
@@ -7,7 +7,7 @@
     <ListboxLabel class="block text-sm font-medium leading-6 text-gray-900">
       {{ label }}
     </ListboxLabel>
-    <div class="relative mt-2">
+    <div class="relative mt-1">
       <ListboxButton class="relative w-full cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 sm:text-sm sm:leading-6">
         <span class="block truncate">{{ selectedOption?.text || placeholder }}</span>
         <span class="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
